[Oztechan/CCC#3957] Move MainView logic into application
